+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.2].define(version: 2025_06_05_122750) do
+ActiveRecord::Schema[7.2].define(version: 2025_10_16_145838) do
   create_table "api_keys", id: :uuid, default: -> { "newid()" }, force: :cascade do |t|
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
@@ -515,6 +515,7 @@ ActiveRecord::Schema[7.2].define(version: 2025_06_05_122750) do
     t.boolean "manage_conversion_urns", default: false
     t.boolean "manage_local_authorities", default: false
     t.datetime "latest_session"
+    t.index ["active_directory_user_id"], name: "UX_users_active_directory_user_id_unique_if_not_null", unique: true, where: "([active_directory_user_id] IS NOT NULL)"
     t.index ["email"], name: "index_users_on_email", unique: true
   end